backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-telegram-bot==21.6
django-environ==0.11.2
drf-yasg==1.21.7
